+++ src/main/java/obro1961/chatpatches/config/Config.java
@@ -39,7 +39,7 @@ public class Config {
     public boolean counterCompact = false; public int counterCompactDistance = 0;
     public boolean boundary = true; public String boundaryFormat = "&8[&r$&8]"; public int boundaryColor = 0x55ffff;
     public boolean chatLog = true, chatHidePacket = true; public int chatWidth = 0, chatMaxMessages = 16384; public String chatNameFormat = "<$>";
-    public int shiftChat = 10; public boolean messageDrafting = false, searchDrafting = true, hideSearchButton = false, vanillaClearing = false;
+    public int shiftChat = 10; public boolean messageDrafting = false, onlyInvasiveDrafting = false, searchDrafting = true, hideSearchButton = false, vanillaClearing = false;
     public int copyColor = 0x55ffff; public String copyReplyFormat = "/msg $ ";